ai-gig-insurance/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.3
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
requests==2.31.0
apscheduler==3.10.4
httpx==0.27.0
python-multipart==0.0.9
